it begins
This begins setting up the processor rewrite. It lays out a structure for the repository, based on WillKG's Antennae and the Processor2015 code. It contains an incomplete sketch of the code, containerization, and runnings tests. At the time of commit, I am developing the rule system.

After spending a while tracing the processor2015 code manually and trying to port it at the same time I decided that too much was happening in the abstract, abandoned writing the whole rule_engine at once, and started working towards getting tests working so that pieces could be tested as they were ported.

At this point I'm still documenting the sources of code snippets in comments, though I expect to remove those as we approach a working v0. This also helps with porting functionality changes as they happen in the processor over to this codebase.

Currently this relies on copies of antennae's testlib, util, and a bunch of other functions. All are targets for shared libraries.

The signatures and functionality look awfully monadic. I'm avoiding that bit of rewrite until I've got a working sketch of the functionality, and then I will investigate adopting a monadic or even applicative style.

I should probably have re-ordered this as docker stuff, then a little code, then tests, but I need to stop working shortly and want to push this somewhere. I'll consider reordering in the future before this becomes a master branch.