refactor crash state into its own class
This isolates some of the functionality of the processor into a crash class. The class closes over the state representing a crash, and provides a transformation method that exposes that state to callables passed as an argument. Some sugar is provided for fetch and save methods that are actually transforms with atypical rules. Those rules are undefined for now.

This also clarifies the Rule object. It has been simplifies to a callable but in the base class retains some backwards compat with processor2016 and before by calling predicate and action methods under the hood. This is intended to make porting old rules easier. It forgoes the version and act methods, as well as the internal _predicate and _action methods that served mostly as error catchers. That responsibility is now put fully on the transform step of the crash class. Error handling is now also in the transform step. I suspect this abstraction will also help deal with filesystem IO when it comes time to implement the breakpad steps, but that is still unclear. Some references remain to the socorro module paths where implementations were speculatively simplified to python builtins and may need to be replaced with more complicated objects.

This change should help clarify the top level Processor class during implementation later. The rule engine class is still an unusable mess and there are probably more things that could be lifted out of antennae that these two projects hold in common.