leverage the test system
enhances testlib with `_` as a throwaway var placeholder and steals the latest loggingmock fixture from antenna. establishes tests for crash and rule objects, including specific tests for specific rules. establishes basic patterns for testing actual rules instead of imagined use cases.

fixes documentation in rule.py to reflect test and actual usage instead of indirect usage described through Crash methods, which are already documented in Crash docstring and in code. Add an Introspector rule for assistance in debugging when writing tests, and the first real transform real extracted from previously unspecc'd processor transform behavior.

crash.py was enhanced with a sugar method for pipelining based on feedback from willkg and adrian, as well as experience starting to use the rule and crash code. this revision also adds an incredible amount of dead code lifted directly from processor2015. this is the first step towards refactoring and integration into this processor rewrite, and was necessary for identifying the first real behavior to incorporate as rules.

finally the parent app was modified to reflect anticipated usage under the current crash object api. processor2015 methods that are no longer applicable to this top-level object were removed and inserted as dead code into the crash module (see above)