add rules for handling processor metadata
processor2015 passed metadata around in explicit objects, every modification of which is cataloged in the CreateMetadata docstring. This integrates handling metadata into the rules system by changing two behaviors from processor2015:

First, duplicate fields being maintained for backwards compatability were removed, as were fields that were set but never read or persisted.

Second, metadata is not stored on the processed_crash object in a complex field during processing and removed after being transformed into top-level fields on the processed_crash object.

With the dead fields removed, the metadata object exists for passing around a freeform notes array, which is a sort of log of what transformed the crash. This could ultimately be simplified and removed if we change the the expected behavior of reprocessing --

Notes from previous processing runs are appended to the end of each other, creating an ordering like:

    C1
    C2
    B1
    B2
    B3
    A1
    A2

This can be intuitive to read back. In the current implementation it is achieved by holding the previous logs in an 'old_notes' variable and later extending the notes from the current run with 'old_notes'. We could remove metadata altogether if we made the notes field append only:

    A1
    A2
    B1
    B2
    B3
    C1
    C2

This behaves like a normal logfile, but when read directly it puts the most important information last. Consider also prepending newlines (or performance optimized equivalents):

    C2
    C1
    B2
    B1
    A2
    A1

Choosing either of these alternatives would remove the overhead of the metadata object creation and deletion.

While considering these optimizations, it's also becoming clear that as more processor2015 behavior is ported over to this processor there will be an opportunity to refactor and combine the builtin rules into some logical phases, such as loading and normalization. Trying not to start that while porting, though. It's too much implied information between commits, and too much for the archeologists looking backwards from tomorrow.