port ESRVersionRewrite and tests
this change brings the TestESRVerisonRewrite and associated tests over from Processor2015. With the processor_notes question unresolved, this change puts the error message into a thrown KeyError instead of in the processor_notes. This makes sense, if you squint, given the changes we've made to the application context of rules. No matter how processor_notes is resolved, this pattern should be forward compatible.