port FennecBetaError20150430
FennecBetaError20150430 has no associated tests in the socorro code base and at this point we could safely remove it from the application. Including for completeness, but not writing new tests.

observational note: these early rules perform many transformations on raw_crashes. In the future we should consider only transforming the processed crash and leaving the raw_crash immutable. Needs investigation and a follow on bug... I'm pretty sure it came up in a recent socorro meeting.