port PluginRule
this change ports PluginRule and associated tests from Process2015, with much simplified logic

archeology points to this particular rule and logic coming in with the rules system. It had a comment with a specific bug number in it that showed the logic was extant and being patched more than 6 years ago in SVN. It was very difficult to follow, with many type conversions and odd logic. This version is much simpler, and could probably be made simpler still. The behavior is subtle and key-names are poorly chosen -- for example, non-hangs may not have a ProcessType.

Running a supersearch for Firefox, all versions, from the last week: only 0.01 percent (about 1k) of crashes are hangs, the rest (1.1 million) are normal crashes. Of the hangs, all but 3 of them are plugin hangs.

Because of that, this change also enhances the test suite with the typical case of a non-hang crash, ensuring that it gets the expected id and hang_type.