port AddonsRule
this change ports AddonsRule and associated tests from Processor2015 with lightly simplified logic

addons are parsed out of a simply formatted string in the raw crash. this inlines the previously static parse function, adds tests for malformed entries, and updates urllib references for python 3.

this also changes the basic structure from try/catch to if/else with early returns

as with previous rule ports, this removes checks for verbose or chatty in the config, though it retains processor_notes since those were checked in the tests.