add time utils
this directly ports datetime, time, and utils more or less directly. I did not spend time figuring out which functions are strictly necessary -- they build on one another and are pretty entagled.

deferring this investigation for now. it will be no more difficult to sort out later, and if anything isolation will make this easier to refactor