port DatesAndTimesRule
Ports over the DatesAndTimesRule and associated tests. Several corrections were necessary for python2->python3: a comparison between None and Int, except as, and comparison of a string to a basestring had to be updated. Additionally a comparison to maxint was replaced with a comparison to maxsize which should function the same way for consistency.

lots of tests ported here, lots of overlap between them. it ended up being reassuring since the logic of this rule was complex and I simplified out a static function and some of the nested `try` blocks.