switch to fixture processed_crash from empty dicts
This switches all uses of empty dicts to fixtures of processed_crash. This is consistent, but perhaps not useful. The fixture is overwritten and then compared against magic strings. The more useful approach is probably to compare transformations captured in an empty dict to the expected outputs in the fixture.