Print help as part of default make target
Every time I run Make I'm greeted with the semihelpful reminder that I need to specify a subcommand. Then I have to go look in the make file to determine which commands are acceptable.

This prints the help command as part of the default message, and adds a documentation line about the help subcommand.